Update test script to boot and kill redis server

+++ script/test.ps1
@@ -5,6 +5,9 @@ function Start-AzureFunction ([int]$port, [string]$workingDir) {
 Start-AzureFunction 7072 -workingDir "test\Indigo.Functions.Configuration.IntegrationTests.Target\bin\Debug\netstandard2.0"
 Start-AzureFunction 7073 -workingDir "test\Indigo.Functions.Injection.IntegrationTests.CorrectConfig\bin\Debug\netstandard2.0"
 Start-AzureFunction 7074 -workingDir "test\Indigo.Functions.Injection.IntegrationTests.NonPublicConfig\bin\Debug\netstandard2.0"
+Start-AzureFunction 7075 -workingDir "test\Indigo.Functions.Redis.IntegrationTests.Target\bin\Debug\netstandard2.0"
+
+redis-server
 
 dotnet test test\Indigo.Functions.Configuration.IntegrationTests\Indigo.Functions.Configuration.IntegrationTests.csproj
 
@@ -18,4 +21,13 @@ if ($LastExitCode -ne 0) {
     return $LastExitCode
 }
 
+dotnet test test\Indigo.Functions.Redis.IntegrationTests\Indigo.Functions.Redis.IntegrationTests.csproj
+
+if ($LastExitCode -ne 0) {
+    return $LastExitCode
+}
+
+# Kill processes only if previous actions succeeded, in case you need to debug
+
 Stop-Process (Get-Process func).Id
+Stop-Process (Get-Process redis-server).Id
